v1: use multiprocessing by default
Previously, this code forced the use of the `spawn` multiprocessing
method. Since we know this causes problems in some configurations,
multiprocessing as off by default.

This change turns it on by default and makes use of existing code that
tries to choose the best multiprocessing method based on what we can
detect.

- use `fork` by default
- use `spawn` if CUDA has already been initialized, but give a warning

This same logic is already in use for spawning multiple workers for v1
tensor parallelism support.

The design doc `docs/design/multiprocessing.md` covers this topic in
more detail.

Signed-off-by: Russell Bryant <rbryant@redhat.com>