v1: Use atexit to handle engine core client shutdown
While testing V1 + multiprocessing, I observed that my offline
inference test script would hang at the end and not exit. There was
shutdown code here that depended on garbage collection to trigger it,
but it wouldn't get far enough for that to happen.

This approach of using `atexit` to register the cleanup handler is the
same as is used in the `multiproc_executor`.

Signed-off-by: Russell Bryant <rbryant@redhat.com>