Improve docs to how to fix multiprocessing conflict
Add info to the debugging tips doc on how  to update the code to use a
`__main__` guard to avoid conflicts with vllm's use of `spawn`.

Update the log message to include a link to the debugging tips page.

Update the more detailed design doc with a reference to the debugging
tips page, as well.

Signed-off-by: Russell Bryant <rbryant@redhat.com>