[Docs] hint to enable use of GPU performance counters in profiling tools
In case of multi-node distributed serving, there is a a helper [script](https://github.com/vllm-project/vllm/tree/main/examples/run_cluster.sh) to start the cluster.  However, this script understandably launches docker without administrative privileges and at times is it curtails the GPU profiling and tracing. The change in the document is to help the users know that if they want to use profiling tools (e.g. NVIDIA NSIGHT) in the docker, they can use `CAP_SYS_ADMIN` to the docker container by using the `--cap-add` option in the docker run command.