Adding max_new_tokens support to generation_config.json
ModelConfig.get_diff_sampling_params() now allows for reading the "max_new_tokens"
if its specified in the generation_config.json file. This follows Huggingface's
naming convention for the variable that specifies the maximum number of generated tokens.
This gets renamed to "max_tokens" to follow the naming convention used by vLLM for the same functionality.

Signed-off-by: Matthew Hendrey <matthew.hendrey@gmail.com>