[Core] Add a feature compatibility checking interface
This change introduces an interface for checking for compatibility
between a set of features that are in use. It uses bitwise math for
checking to be able to perform the check in fewer operations than other
potential approaches.

This changle also makes use of the interface in the v0 engine to detect
the conflict between speculative decoding and structured output (guided
decoding). Without this in place, vllm will crash later on when trying
to process the request.

Signed-off-by: Russell Bryant <rbryant@redhat.com>