Add SPDX license headers to python source files
This commit adds SPDX license headers to python source files as recommended to
the project by the Linux Foundation. These headers provide a concise way that is
both human and machine readable for communicating license information for each
source file. It helps avoid any ambiguity about the license of the code and can
also be easily used by tools to help manage license compliance.

The Linux Foundation runs license scans against the codebase to help ensure
we are in compliance with the licenses of the code we use, including
dependencies. Having these headers in place helps that tool do its job.

More information can be found on the SPDX site:

- https://spdx.dev/learn/handling-license-info/

Signed-off-by: Russell Bryant <rbryant@redhat.com>