[Core] Don't do platform detection at import time
Running just `vllm` or `vllm --version` or something similar resulted
in several lines of log noise about platform detection. This is because
several places were triggering this to occur at import time by doing the
following:

```python
from vllm.platforms import current_platform
```

Platform detection runs the first time the `current_platform` attribute is requested, which includes this import. Importing `vllm.platforms` and delaying the reference to `current_platform` until code actually needs it avoids running it in cases where it was never needed.

I only changed the places that caused a problem in my environment when running `vllm`, though I'd say importing it like this at the top of a module is an anti-pattern in general.

Signed-off-by: Russell Bryant <rbryant@redhat.com>