[V1][Core] Generic mechanism for handling engine utility methods
We now have a number of utility / "control" operations that need to be called on the engine (add_lora, profile, sleep, wakeup, ...). It should be possible to call these in a synchronous manner to know when the operation is complete and whether it succeeded. Some operations in future may require results to be returned.

These changes centralize the mechanism for doing this in V1, and add the result-returning part.

Signed-off-by: Nick Hill <nhill@redhat.com>