[Frontend] Pass pre-created socket to uvicorn
I noticed that the `fd=...` argument had been re-added to our uvicorn
configuration, but only for mac, in PR #11696.. A previous PR, #10012,
includes an explanation of why we should not be using this argument,
as it does not behave as intended.

There is another way to do what we really want, which is to have
uvicorn re-use a socket we have already created. This change
implements that.

For the pre-created socket, we previously set `SO_REUSEADDR`, but this
change adds `SO_REUSEPORT` as well. While it's not strictly necessary
in this PR, it will be needed when we start running multiple API
server processes (issue #12705). I was already changing related code
here, and the change also helps demonstrate the value of re-using the
existing socket. A helpful explanation of `SO_REUSEPORT` can be found
here: <https://lwn.net/Articles/542629/>

One side-effect of this change is that uvicorn no longer emits a
message showing the host and port number in use when it starts the
server, so this includes a new INFO log message on the vLLM side that
does the same thing prior to starting the server.

Signed-off-by: Russell Bryant <rbryant@redhat.com>