[Misc] Move pre-commit suggestion back to the end
This change moves the pre-commit suggestion back to the end of the
pre-commit configuration. This entry emits a message about how to skip
the pre-commit checks. Its better to have that at the end instead of the
middle.

I also included some comments to remind future editors of the file to keep this
entry at the end.

Signed-off-by: Russell Bryant <rbryant@redhat.com>