[CI/Build] Ignore ruff warning up007
This change ignores the following warning from ruff:

  UP007 Use `X | Y` for type annotations

We need to continue using the `Optiona[T]` syntax until Python 3.10
or above is our minimum supported version. I'm not sure why I only
see this in one of my environments, but not another. In any case,
it seems harmless to explicitly ignore it.

Signed-off-by: Russell Bryant <rbryant@redhat.com>