Changes to get a test request working
In the scheduler, if a request is still waiting on FSM completion,
we now push it back to the end of the waiting queue.

Each time we check on a request in this state, we also check to see
if it has actually finished. It was previously stuck without ever
getting the result.

Signed-off-by: Russell Bryant <rbryant@redhat.com>