[V1][Metrics] Implement max_num_generation_tokens metrics
This metric tracks the maximum of num_generation_tokens across
a set of identical requests under a parallel sampling parent.

It is the last remaining metric used by the example Grafana
dashboard that makes sense in V1.

Add some additional tracking of child requests to ParentRequest
in order to facilitate this.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>