[V1] Simplify stats logging
Currently the core engine loop wakes up every 2.5 seconds when idle to return scheduler stats to the client, but these won't have changed if there are no unfinished requests.

Update the StatsLogger interface to have record() and log() methods and wire the latter to the existing stats logging task in the api server rather than relying on regular updates even if there's no change. This helps to simplify the engine core loop.

I don't think responsibility for the periodic log_stats task should be in the API server, we should move it inside the engine client and remove the external EnigneClient.do_log_stats method. But that can be changed in a subsequent PR.

Signed-off-by: Nick Hill <nhill@redhat.com>