[V1] Eagerly remove finished requests from the batch
Currently, if all in-flight requests are aborted, the core engine loop will pause and only remove the completed requests from the batch the next time there are new requests to schedule.

We might as well do this early in this case so that it's not on the later critical path.

Other advantages of this:
- It makes it easy to return a final CoreRequestOutputs with stats that reflect that all requests are finished, so that they can be updated without having to poll the engine.
- It will help with the data parallel implementation where we need to coordinate when to pause multiple engines based on there being no in-flight requests.

Signed-off-by: Nick Hill <nhill@redhat.com>