[Bugfix]: Fix Promethus spec decode counter sum-of-sums
The Prometheus spec decode counters (draft/accepted/emitted token counts) are incremented by the values in spec_decode_metrics. However, those values are aggregates since startup. Therefore, the Prometheus counters are effectively a sum-of-sums instead of just a sum.

If a high-traffic vLLM is left on for a few hours those counters start to suggest absurdly high values like a TPS in the tens of millions.

Signed-off-by: Adam Lugowski <adam.lugowski@parasail.io>