[V1] Delay all xgrammar usage until needed
PR #14575 delayed initialization of the grammar bitmask until it was
needed to try to fix a problem encountered on TPU systems.
Unfortunately, that change was not sufficient.

We need to delay usage of ALL xgrammar APIs, not just the grammar
initialization. This change implements that. More initialization is now
deferred until the first time a structured output request is received.

Signed-off-by: Russell Bryant <rbryant@redhat.com>