[V1] Refactor Structured Output for multiple backends
This change does some refactoring of the V1 structured output
implementation to prepare for supporting multiple backends. This code is
already successfully in use in a branch to support a second backend. I
think it will be easier to review other backends if the refactoring goes
in first on its own.

Signed-off-by: Russell Bryant <rbryant@redhat.com>