setup.py: drop assumption about local `main` branch
PR #13921 added an assumption about the local `main` branch. We're
already getting the latest commit ref for `main` upstream, so using this
branch is not necessary in the rest of the logic. What we really care
about is upstream main and the git branch we're currently on.

This change should keep the same behavior, but stop looking at the
`main` branch as part of it.

Signed-off-by: Russell Bryant <rbryant@redhat.com>