Disable outlines cache by default
https://github.com/vllm-project/vllm/security/advisories/GHSA-mgrm-fgjv-mhv8

Outlines provides a cache for its compiled grammars on the local
filesystem. This cache has been on by default in vLLM. Outlines is also
available by default through the OpenAI compatible API server.

A malicious user can send a stream of very short decoding requests with
unique schemas, resulting in an addition to the cache for each request.
This can result in a Denial of Service if the filesystem runs out of
space.

Note that even if vLLM was configured to use a different backend by
default, it is still possible to choose outlines on a per-request basis
using the `guided_decoding_backend` key of the `extra_body` field of the
request.

This issue applies to the V0 engine only. The V1 engine is not affected.

Signed-off-by: Russell Bryant <rbryant@redhat.com>