[BugFix] Lazily import XgrammarBackend to avoid early cuda init
Importing xgrammar appears to initialize the cuda context, which we don't want to do in the front-end process. It also means that the server can't be started with the (default) multiproc context mode of fork.

I guess this is what LazyLoader is meant to help with, but it doesn't seem to be working as intended since https://github.com/vllm-project/vllm/pull/14694 was merged.

Signed-off-by: Nick Hill <nhill@redhat.com>