[SW-218309] Updated Tests Workflow To Support Forked PR's (#930)
Changes:

- Changed the trigger of the tests from pull_request to issue_comment.
now tests will be triggered only if a CODEOWNER will comment on the PR
/run-gaudi-tests
- Add random string to hlctl container when running the tests. I noticed
that a lot of times in case of re-runs , tests is failing due to the
error of workload already exists. the random string will solve it
- Added a job that will collect all tests status and will fail/pass
based on the tests execution status. It will be used as a mandatory
status check in order to merge a PR(means that without all tests passed,
a PR cannot be merged)