[BugFix][V1] Quick fix for min_tokens with multiple EOS
Reported by @andylolu2 in slack. This can affect min_tokens function for models that define more than one EOS token in their config.

We need to revisit how SamplingParams.__post_init__ is used generally, as well as test coverage for this specific case. This is a quick fix in the meantime for a point release planned for today.

Signed-off-by: Nick Hill <nhill@redhat.com>