[V1] Fix json_object support with xgrammar
Our previous behavior with xgrammar was to ensure that the response was
valid JSON. When the json_object output format is requested, the correct
behavior is to ensure the response is a JSON object. Valid JSON could
also be an array, a string, or a number, which is not desired here.

The fix is to explicitly use a json schema of {"type": "object"}. An
upgrade to xgrammar is necessary to fix a bug with xgrammar's support of
this schema, so we upgrade to 0.1.17.

- https://github.com/mlc-ai/xgrammar/issues/256
- https://github.com/mlc-ai/xgrammar/pull/264

Signed-off-by: Russell Bryant <rbryant@redhat.com>