[V1] Update structured output offline inference example
Update this example to work with V1. The grammar specified was not
comaptible with xgrammar. This worked with V0 as we would automatically
fall back to outlines.

In V1 we have fallback support, but you have to opt-in to it by setting
your backend to `auto`. Grammar type support varies among the backends.

Signed-off-by: Russell Bryant <rbryant@redhat.com>