Update deepseek modeling file for PatchedVLLMKVCache (#1009)
Previously, when we use INC to convert deepseek FP8 model, we need this
[commit
](https://github.com/intel/neural-compressor/commit/7c0a3e20f48e89707c7be0ec0e4719e6bbc55bac)
to remove extra converts in KVCache but actually GC can remove them
during graph optimization theoretically.
Furthermore, the change in commit is not aligned with the design of INC
patched module, which wants to keep the returned tensor BF16 because we
can't make sure users' next operation.
So, I update the modeling file to make GC can work for patched KVCache
pattern of deepseek model.
Since next release is very close and GC currently can not work as
expection during decode stage, it is still a workround. We will root
cause and fix it from source in next relase.

This PR should work together with this PR:
https://github.com/intel/neural-compressor/pull/2165

Signed-off-by: Mengni Wang <mengni.wang@intel.com>