add ScaleToHwAligned for loading fp8 vllm model (#941)
https://jira.habana-labs.com/browse/SW-207506 the scales provided by
neuralmagic fp8 model method are maxabs, it need do hw scale align to
adapt hpu platform to get better accuracy and performance. I add the
class `ConvertScaleHwAlign` in
vllm-hpu-extension(https://github.com/HabanaAI/vllm-hpu-extension/pull/118)
and call it in vllm-fork. the class also include the device check and
make factor first if the device is G2.
it is used for loading the models in this link.
https://huggingface.co/collections/neuralmagic/fp8-llms-for-vllm-666742ed2b78b7ac8df13127

---------

Signed-off-by: changwangss <changwang@habana.ai>