[V1][DP] More robust DP/EP dummy request coordination
The current mechanism for coordinating dummy forward passes between EP ranks is not very intuitive and there may be a race condition.

This changes to keep track of a sequence number for request "waves" which should be more robust and reduce the number of control messages that need to be sent. Also rename the engine core "global_unfinished_reqs" field to "engines_running" which is more intuitive and consistent with equivalent flag on the client side.

Signed-off-by: Nick Hill <nhill@redhat.com>