[Core] Upgrade to xgrammar 0.1.18, add cache size limit
xgrammar 0.1.18 includes a fix for security vulnerability where a
malicious user can cause vllm to consume all of the host's RAM with
the knowledge that each unique grammar adds an entry to an unbounded
cache in memory.

The cache was implemented in xgrammar PR 243:
https://github.com/mlc-ai/xgrammar/pull/243

This commit makes use of the new option to limit the cache size.
The default is 512MB, which should be reasonable for most use cases.
It will cache roughly 1000 json schemas. The cache size can be changed
using the `VLLM_XGRAMMAR_CACHE_MB` environment variable.

The security vulnerability was reported here:
https://github.com/mlc-ai/xgrammar/security/advisories/GHSA-389x-67px-mjg3

Signed-off-by: Russell Bryant <rbryant@redhat.com>