Fix tests to be more V1 friendly
These tests now run against V1 in CI, but the code was originally
written with V0 in mind. In particular, specifying the backend
through the OpenAI API is not supported with V1, so we can remove
it and speed up the tests quite a bit. Different backends are tested in
another place (via the llm entrypoint tests).

Signed-off-by: Russell Bryant <rbryant@redhat.com>