[DeepSeek R1] Skip profile run on decode server (#1071)
In PD disaggregation scenario, it doesn't make sense to profile run with
the largest prefill bucket on the decode instances. However, it can not
run with decode batch because the number of hpu blocks is unknown at
this stage. The PR skip the profile run on decode instances and it is
expected to use VLLM_GPU_MEMORY_UTILIZATION to reserve enough memory.