Add option to use torch._inductor.standalone_compile
This PR adds the option to use torch._inductor.standalone_compile to
perform compilation instead of compile_fx. The goal of
standalone_compile is to remove the hacks around vLLM's usage of
compile_fx, we want to migrate to using it in PyTorch 2.8.

standalone_compile replaces how vLLM interacts with the torch.compile
caches. Instead of vLLM trying to redirect them into its
torch_compile_cache folder, vLLM can pass standalone_compile a filepath
that is inside of the torch_compile_cache folder and standalone_compile
will write the full precompiled artifact to it.

Right now this option is hidden behind an envvar. It is also not
tested in vLLM CI (vLLM CI only tests against PyTorch 2.6).
This option also needs more testing before we turn it on by default for
PyTorch 2.8+. I am putting this PR out so that we can
merge something that we can keep developing on top of.

Test Plan:
- Run https://gist.github.com/zou3519/aebb622714e80f4cd4c369472f2372cd
with or without VLLM_TEST_STANDALONE_COMPILE

Signed-off-by: rzou <zou3519@gmail.com>