multi-image support for llama3.2 [1/N] (#926)
This PR is identical with #705 , except that it's from a branch within
vllm-fork instead of a branch from an external forked repo.

The reason for this is that there are some issues for merge external
branch, see jira
[SW-218309](https://jira.habana-labs.com/browse/SW-218309), we have
consulted folks in SW-218309, they suggest to create a internal branch
to trigger the jenkins test correctly, so I am creating this PR to
speedup the merge process.

cc @yma11 @michalkuligowski @jkaniecki

---------

Signed-off-by: yan ma <yan.ma@intel.com>
Signed-off-by: zhouyu5 <yu.zhou@intel.com>
Co-authored-by: yan ma <yan.ma@intel.com>
Co-authored-by: Michał Kuligowski <mkuligowski@habana.ai>