[V1] Allow turning off pickle fallback in vllm.v1.serial_utils
`pickle` is a well-known insecure method of serialization if there is
any chance we may be deserializing untrusted data. It is the cause of
many security vulnerabilities, including ones in vLLM. The move to use
`msgpack` instead of `pickle` throughout the codebase is a significant
improvement, though our custom encoder/decoder includes a fallback to
`pickle` when we don't know else to handle a given type.

Add a new option to `vllm.v1.serial_utils` which allows turning off the
`pickle` fallback. We should try to set `allow_pickle=False` anywhere
that serialization is being used over a network, in particular.

Use of this option will come in later PRs.

Signed-off-by: Russell Bryant <rbryant@redhat.com>