[easy] Fix logspam on PiecewiseBackend errors
Sometimes when there is an error, we get spammed with a dump of the
current torch.fx.Graph that is being compiled. That is due to this code
here:
https://github.com/pytorch/pytorch/blob/fc6e37ceb23f99808265c11a37368078d5f982b8/torch/fx/interpreter.py#L173-L183

We turn off this flag so we don't get logspams. The spam is not helpful,
also vLLM has other ways of dumping the graphs to file.

Signed-off-by: rzou <zou3519@gmail.com>