[Core] Remove prompt string from engine core data structures
By design, V1 engine core deals only with token ids. EngineCoreRequest contained an optional string prompt field, but we always set it to None before sending to the engine. It's used only by the output processor to set the prompt back in output messages returned to the user.

It's best to remove it from EngineCoreRequest to make this separation clear/explicit.

Signed-off-by: Nick Hill <nhill@redhat.com>