Ignore `'<string>'` filepath
vLLM tries to get a list of files that Dynamo traced through. Some of
these files can be equal to `'<string>'`. This may happen if the code
object was dynamically generated, which does happen in some dataclasses
cases.

This is one of the issues that was blocking the upgrade of vLLM from
PyTorch 2.6 to PyTorch 2.7 (#16859). The reason why this happened was because
PyTorch 2.7 now traces through more HuggingFace code, including some
custom dataclasses.

Test Plan:
- ran `pytest -v tests/models/test_transformers.py -k test_models[meta-llama/Llama-3.2-1B-Instruct-transformers]` with PyTorch >= 2.7

Signed-off-by: rzou <zou3519@gmail.com>