Initialize the delta tool call fields explicitly
Previously the implementation relied on a combination
of the default field values being and
`model_dump_json(exclude_unset=True)`. This led to bugs such
as the "type" field missing from the response.

The behavior of the OpenAI API is to set the "id" and "type"
fields only in the first Delta object.

The purpose of this commit is to make all the places
where these fields need to be set explicit.

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>
Co-authored-by: igmainc <igmainc@icloud.com>