[Deepseek] Enable INC WoQ Requant for G2 (#1153)
The [Deepseek R1](https://huggingface.co/deepseek-ai/DeepSeek-R1), V3,
and V3-0324 include the FP8 weights with block-wise scales. Previously,
the
[WoQ](https://github.com/HabanaAI/vllm-fork/blob/deepseek_r1/scripts/DEEPSEEK_R1_ON_GAUDI.md#option-1-run-with-runtime-dequantize-with-block-based-scale)
was supported, which loading the original FP8 weight and block-wise
scales, performs on-the-fly dequantization, and uses BF16 GEMM for
computation.

Building on this approach, this PR adds support for re-quantizing FP8
weights to a format with a INC supported granularity—either per-tensor
or per-channel.

Specifically:
- At calibration stage, it leverage the WoQ to measure the model as a
BF16 model.
- During convert stage, it first dequantize the weight w/ block-wise
scales to BF16, and then re-quantize weights to FP8 using either
per-channel or per-tensor scaling.

For G3, we can directly use the official FP8 weights. For G2, however,
the weights must first be normalized using the `convert_for_g2.py`
script.

cc @thuang6

---------

Signed-off-by: yiliu30 <yi4.liu@intel.com>
Signed-off-by: Yi <yi4.liu@intel.com>
Signed-off-by: Yi Liu <yiliu4@habana.ai>
Signed-off-by: yi <yi>
Co-authored-by: Yi Liu <yiliu4@habana.ai>
Co-authored-by: yi <yi>
Co-authored-by: Youlei Yang <youlei.yang@intel.com>