[V1] Disable pickle by deafult for new serial_utils usage
This changes vllm.v1.serial_utils to disable the pickle fallback
behavior by default for new code using these interfaces. All existing
usage has been updated to use `allow_pickle=True` so that nothing
breaks.

This is a better default, as it requires code authors to explicitly
acknowledge and declare that `pickle` is in use, which must be analyzed
for security risk before turned on.

Signed-off-by: Russell Bryant <rbryant@redhat.com>